Refactor almost everything about blocks and block states
While I was at it I also created a one-time pixlyzer reformatter so
that on any launch except the first launch, the registries can just
simply be loaded from json. I may even be able to use sticky-encoding
instead of json coding to further speed up registry loading. Fixes #25.
Fixes #21 (waterlogged block rendering). Fluids are now a lot nicer to
work with as well.

The main changes made to block stuff is that there is no such thing as
block states in Delta Client now. Each 'block state' is just its own
block. This reduces the amount of lookups and really centralises
everything nicely.

+++ Sources/DeltaCore/World/Event/SetBlock.swift
@@ -3,6 +3,6 @@ import Foundation
 extension World.Event {
   struct SetBlock: Event {
     let position: Position
-    let newState: UInt16
+    let newState: Int
   }
 }
